Move post_upgrade method to generators folder
