+++ kybra/compiler/kybra_generate/src/generators/canister_methods/mod.rs
@@ -1,4 +1,5 @@
 pub mod heartbeat;
 pub mod init;
 pub mod inspect_message;
+pub mod post_upgrade;
 pub mod pre_upgrade;
